+++ scilab/modules/gui/src/cpp/GetUicontrolPosition.cpp
@@ -18,7 +18,7 @@ using namespace org_scilab_modules_gui_bridge;
 int GetUicontrolPosition(sciPointObj* sciObj)
 {
   int returnFlag = FALSE;
-  long int * returnValues = NULL;
+  int * returnValues = NULL;
   double * tmp = NULL;
 
   if (sciGetEntityType( sciObj ) == SCI_UICONTROL)
