+++ scilab/modules/renderer/src/cpp/subwinDrawing/IsoViewCameraJavaMapper.cpp
@@ -109,7 +109,7 @@ void IsoViewCameraJavaMapper::setAxesReverse(bool xAxisRevert, bool yAxisRevert,
 void IsoViewCameraJavaMapper::getPixelCoordinates(double userCoordX, double userCoordY, double userCoordZ,
                                             int pixCoord[2])
 {
-  long * javaCoords = m_pJavaObject->get2dViewPixelCoordinates(userCoordX, userCoordY, userCoordZ);
+  int * javaCoords = m_pJavaObject->get2dViewPixelCoordinates(userCoordX, userCoordY, userCoordZ);
   pixCoord[0] = javaCoords[0];
   pixCoord[1] = javaCoords[1];
 
@@ -119,7 +119,7 @@ void IsoViewCameraJavaMapper::getPixelCoordinates(double userCoordX, double user
 void IsoViewCameraJavaMapper::get2dViewPixelCoordinates(double userCoordX, double userCoordY, double userCoordZ,
                                            int pixCoord[2])
 {
-  long * javaCoords = m_pJavaObject->get2dViewPixelCoordinates(userCoordX, userCoordY, userCoordZ);
+  int * javaCoords = m_pJavaObject->get2dViewPixelCoordinates(userCoordX, userCoordY, userCoordZ);
   pixCoord[0] = javaCoords[0];
   pixCoord[1] = javaCoords[1];
